+++ schemars/src/json_schema_impls/core.rs
@@ -35,8 +35,7 @@ impl<T: JsonSchema> JsonSchema for Option<T> {
             }
         }
         if gen.settings().option_nullable {
-            let mut schema_obj = schema.into();
-            gen.make_extensible(&mut schema_obj);
+            let mut schema_obj = schema.into_object();
             schema_obj
                 .extensions
                 .insert("nullable".to_owned(), json!(true));
@@ -178,8 +177,7 @@ forward_impl!((<'a> JsonSchema for std::fmt::Arguments<'a>) => String);
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::gen::*;
-    use crate::tests::{custom_schema_object_for, schema_for, schema_object_for};
+    use crate::tests::{schema_for, schema_object_for};
     use pretty_assertions::assert_eq;
 
     #[test]
@@ -207,21 +205,6 @@ mod tests {
         assert_eq!(any_of.len(), 2);
         assert_eq!(any_of[0], Schema::new_ref("#/definitions/Foo".to_string()));
         assert_eq!(any_of[1], schema_for::<()>());
-    }
-
-    #[test]
-    fn schema_for_option_with_nullable() {
-        let settings = SchemaSettings::default().with(|s| {
-            s.option_nullable = true;
-            s.option_add_null_type = false;
-        });
-        let schema = custom_schema_object_for::<Option<i32>>(settings);
-        assert_eq!(
-            schema.instance_type,
-            Some(SingleOrVec::from(InstanceType::Integer))
-        );
-        assert_eq!(schema.extensions.get("nullable"), Some(&json!(true)));
-        assert_eq!(schema.subschemas.is_none(), true);
     }
 
     #[test]